corrected typo of exiu-ccd-connector-wrapper
